refactor: 重构评分系统并优化数据库迁移
- 将评分字段从固定值改为范围验证(0.5-5.0)
- 添加评分评论字段
- 优化数据库迁移脚本，修复索引创建语法
- 简化BoxOffice数据结构，移除冗余字段
- 重构评分服务逻辑，简化提交流程
- 添加全面的测试脚本(简单测试、详细测试、端到端测试)
- 更新依赖项并修复迁移过程中的脏状态处理
- 修改默认端口为9090并强制使用本地数据库连接

+++ internal/migrations/000002_create_ratings_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS ratings (
     movie_title VARCHAR(255) NOT NULL,
     rater_id VARCHAR(255) NOT NULL,
-    rating NUMERIC(2,1) NOT NULL CHECK (rating IN (0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0)),
+    rating FLOAT NOT NULL CHECK (rating IN (0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0)),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (movie_title, rater_id),
